+++ Sources/EmojiKit/Emoji+All.swift
@@ -10,17 +10,23 @@ import SwiftUI
 
 public extension Emoji {
 
-    /// Get all available emojis from all categories.
-    static var all: [Emoji] {
-        EmojiCategory.standard.flatMap {
+    /// Get all emojis from all standard categories.
+    ///
+    /// This property only returns emojis that are available
+    /// the current runtime.
+    static let all: [Emoji] = {
+        EmojiCategory.standardCategories.flatMap {
             $0.emojis
         }
-    }
+    }()
 }
 
 public extension Collection where Element == Emoji {
 
     /// Get all available emojis from all categories.
+    ///
+    /// This property only returns emojis that are available
+    /// the current runtime. 
     static var all: [Element] {
         Element.all
     }